+++ contracts/adapters/CurveAdapter.sol
@@ -39,8 +39,8 @@ contract CurveAdapter is Adapter {
      * @return Amount of stableswapToken locked on the protocol by the given user.
      * @dev Implementation of Adapter function.
      */
-    function getAssetAmount(address asset, address user) external view override returns (int128) {
-        return int128(ERC20(asset).balanceOf(user));
+    function getAssetAmount(address asset, address user) external view override returns (int256) {
+        return int256(ERC20(asset).balanceOf(user));
     }
 
     /**
